+++ internal/ui/model.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/Sendspin/sendspin-go/internal/sync"
+	"github.com/Sendspin/sendspin-go/pkg/sync"
 	tea "github.com/charmbracelet/bubbletea"
 )
